Rename "time" module to "serialization"
