+++ src/main.rs
@@ -7,7 +7,7 @@ mod indicators;
 mod markets;
 mod portfolio;
 mod strategies;
-mod time;
+mod serialization;
 mod traits;
 mod types;
 mod engine;
